EVEREST-355-default-database-name (#96)
